examples/compaction/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ariadne-compaction"
version = "0.1.0"
description = "Auto-compaction extension for OpenAI Agents SDK with Ariadne trace viewer integration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"openai>=1.0.0",
	"openai-agents>=0.1.0",
	"tiktoken>=0.7.0",
	"pyyaml>=6.0",
	"python-dotenv>=1.0.0",
	"httpx>=0.27.0",
	"requests>=2.31.0",
]

[project.optional-dependencies]
s3 = ["boto3>=1.28.0"]
dev = [
	"pytest>=7.4.0",
	"pytest-cov>=4.1.0",
	"mypy>=1.5.0",
	"ruff>=0.1.0",
]

[project.scripts]
compactctl = "compact.cli:main"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["compact*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=compact --cov-report=html --cov-report=term-missing"
